week 1 practice day done

+++ prefix_sum_and_binary_search.cpp/prefix_sum.cpp
@@ -4,13 +4,15 @@ using namespace std;
 int main()
 {
 
-  long long n, q;
+  int n, q;
   cin >> n >> q;
   long long a[n];
+
   for (int i = 0; i < n; i++)
   {
     cin >> a[i];
   }
+
   long long pre[n];
   pre[0] = a[0];
   for (int i = 1; i < n; i++)
@@ -24,6 +26,7 @@ int main()
     cin >> l >> r;
     l--;
     r--;
+
     long long sum;
     if (l == 0)
       sum = pre[r];
@@ -31,5 +34,6 @@ int main()
       sum = pre[r] - pre[l - 1];
     cout << sum << endl;
   }
+
   return 0;
 }
